requirements/base.txt
Django==2.1.7
djangorestframework==3.9.1
djangorestframework-jwt==1.11.0
